Massive upgrade to params file read/write
