+++ openbb_terminal/miscellaneous/i18n/en.yml
@@ -491,10 +491,8 @@ en:
   crypto/disc/games: top blockchain games
   crypto/disc/dapps: top decentralized apps
   crypto/disc/dex: top decentralized exchanges
-  crypto/ov/global: global crypto market info
   crypto/ov/defi: global DeFi market info
   crypto/ov/stables: stablecoins
-  crypto/ov/exchanges: top crypto exchanges
   crypto/ov/exrates: coin exchange rates
   crypto/ov/indexes: crypto indexes
   crypto/ov/derivatives: crypto derivatives
@@ -968,7 +966,7 @@ en:
   portfolio/po/_loaded: Portfolio loaded
   portfolio/po/_tickers: Tickers
   portfolio/po/_categories: Categories
-  portfolio/po/file: select portfolio parameter file
+  portfolio/po/params/load: select portfolio parameter file
   portfolio/po/params: specify and show portfolio risk parameters
   portfolio/po/_parameter: Parameter file
   portfolio/po/_mean_risk_optimization_: Mean Risk Optimization
